requirements.txt
aioitertools==0.12.0
aiosqlite==0.20.0
alembic==1.13.3
annotated-types==0.7.0
anyio==4.6.2.post1
arize-phoenix==5.5.2
arize-phoenix-evals==0.17.2
arize-phoenix-otel==0.6.1
Authlib==1.3.2
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.3
Deprecated==1.2.14
distro==1.9.0
fast-hdbscan==0.2.0
fastapi==0.115.3
Flask==3.0.3
Flask-Cors==5.0.0
googleapis-common-protos==1.65.0
graphql-core==3.2.5
greenlet==3.1.1
grpc-interceptor==0.15.4
grpcio==1.67.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
importlib_metadata==8.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.6.1
joblib==1.4.2
llvmlite==0.43.0
Mako==1.3.6
MarkupSafe==3.0.2
numba==0.60.0
numpy==2.0.2
openai==1.52.2
openinference-instrumentation==0.1.18
openinference-instrumentation-openai==0.1.14
openinference-semantic-conventions==0.1.11
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-exporter-otlp-proto-http==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-flask==0.48b0
opentelemetry-instrumentation-requests==0.48b0
opentelemetry-instrumentation-wsgi==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
packaging==24.1
pandas==2.2.3
phoenix==0.9.1
protobuf==4.25.5
psutil==6.1.0
pyarrow==17.0.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pynndescent==0.5.13
python-dateutil==2.9.0.post0
python-multipart==0.0.12
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
scikit-learn==1.5.2
scipy==1.14.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
sqlean.py==3.47.0
starlette==0.41.0
strawberry-graphql==0.243.1
threadpoolctl==3.5.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
umap-learn==0.5.6
urllib3==2.2.3
uvicorn==0.32.0
websockets==13.1
Werkzeug==3.0.6
wrapt==1.16.0
zipp==3.20.2
